refactor: move veto changelistener into util
